ensure datetimes on distributors
closes #1254
