+++ server/pulp/server/managers/repo/_common.py
@@ -23,22 +23,6 @@ from pulp.server import config as pulp_config
 from pulp.server.exceptions import PulpCodedException
 
 
-def _ensure_tz_specified(time_stamp):
-    """
-    Check a datetime that came from the database to ensure it has a timezone specified in UTC
-    Mongo doesn't include the TZ info so if no TZ is set this assumes UTC.
-
-    :param time_stamp: a datetime object to ensure has UTC tzinfo specified
-    :type time_stamp: datetime.datetime
-    :return: The time_stamp with a timezone specified
-    :rtype: datetime.datetime
-    """
-    if time_stamp:
-        time_stamp = dateutils.to_utc_datetime(time_stamp, no_tz_equals_local_tz=False)
-
-    return time_stamp
-
-
 def to_transfer_repo(repo_data):
     """
     Converts the given database representation of a repository into a plugin
@@ -53,8 +37,8 @@ def to_transfer_repo(repo_data):
     """
     r = Repository(repo_data['id'], repo_data['display_name'], repo_data['description'],
                    repo_data['notes'], content_unit_counts=repo_data['content_unit_counts'],
-                   last_unit_added=_ensure_tz_specified(repo_data.get('last_unit_added')),
-                   last_unit_removed=_ensure_tz_specified(repo_data.get('last_unit_removed')))
+                   last_unit_added=dateutils.ensure_tz(repo_data.get('last_unit_added')),
+                   last_unit_removed=dateutils.ensure_tz(repo_data.get('last_unit_removed')))
     return r
